+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 
 - **New:** Seed for thread prefix (#50)
 - **New:** Seed for resource prefix (#51)
+- **New:** Seed for thread prefix group (#52)
 
 ## 1.1.0 Alpha 4 (`1010014`)